Refactor code in main.rb and handle edge case.
Split code into different files and test that all game edge cases are worked on.
